Check Add not found handling

+++ app.py
@@ -28,6 +28,8 @@ def search_item():
 @app.route("/item/<int:item_id>")
 def show_item(item_id):
     item = items.get_item(item_id)
+    if not item:
+        abort(404)
     return render_template("show_item.html", item=item)
 
 @app.route("/new_item")
@@ -49,8 +51,11 @@ def create_item():
 @app.route("/edit_item/<int:item_id>")
 def edit_item(item_id):
     item = items.get_item(item_id)
+    if not item:
+        abort(404)
     if item["user_id"] != session["user_id"]:
         abort(403)
+
     return render_template("edit_item.html", item=item)
 
 @app.route("/update_item", methods=["POST"])
@@ -59,6 +64,7 @@ def update_item():
     item = items.get_item(item_id)
     if item["user_id"] != session["user_id"]:
         abort(403)
+
     book_name = request.form["book_name"]
     author = request.form["author"]
     grade = request.form["grade"]
@@ -71,12 +77,13 @@ def update_item():
 @app.route("/delete_item/<int:item_id>", methods = ["GET", "POST"])
 def delete_item(item_id):
     item = items.get_item(item_id)
+    if not item:
+        abort(404)
     if item["user_id"] != session["user_id"]:
         abort(403)
 
     if request.method == "GET":
         return render_template("delete_item.html", item=item)
-
     if request.method == "POST":
         if "remove" in request.form:
             items.delete_item(item_id)
